# 211 - lex start