--- a/mf2wg_test.go
+++ b/mf2wg_test.go
@@ -22,9 +22,6 @@ func init() {
 	//nolint:lll
 	failing = []string{
 		"TestMF2WG/.message-format-wg/test/tests/functions/datetime.json/{|2006-01-02T15:04:06|_:datetime_year=numeric_month=|2-digit|}",
-
-		"TestMF2WG/.message-format-wg/test/tests/syntax.json/__{{}}__",
-		"TestMF2WG/.message-format-wg/test/tests/syntax.json/_.input{$x}{{}}_",
 	}
 }