--- a/parse/parse_test.go
+++ b/parse/parse_test.go
@@ -802,7 +802,7 @@ func BenchmarkParse(b *testing.B) {
 	var tree AST
 
 	for range b.N {
-		tree, _ = Parse(".input {$foo :number} .local $bar = {$foo} .match {$bar} one {{one}} * {{other}}")
+		tree, _ = Parse("  .input {$foo :number} .local $bar = {$foo} .match {$bar} one {{one}} * {{other}}  ")
 	}
 
 	runtime.KeepAlive(tree)