* lex the start
* handle whitespaces at the beginning and the end

Fixes #211 

```
➜  parse git:(211-lex-start) ✗ benchstat a.txt b.txt
goos: darwin
goarch: arm64
pkg: go.expect.digital/mf2/parse
cpu: Apple M2 Max
                         │    a.txt    │               b.txt                │
                         │   sec/op    │   sec/op     vs base               │
ComplexMessage_String-12   3.078µ ± 1%   3.134µ ± 1%  +1.84% (p=0.000 n=10)
Lex-12                     1.503µ ± 0%   1.581µ ± 1%  +5.19% (p=0.000 n=10)
Parse-12                   3.199µ ± 1%   3.334µ ± 2%  +4.22% (p=0.000 n=10)
geomean                    2.455µ        2.547µ       +3.74%

                         │    a.txt     │                 b.txt                 │
                         │     B/op     │     B/op      vs base                 │
ComplexMessage_String-12   20.33Ki ± 0%   20.33Ki ± 0%       ~ (p=1.000 n=10) ¹
Lex-12                       528.0 ± 0%     528.0 ± 0%       ~ (p=1.000 n=10) ¹
Parse-12                   7.000Ki ± 0%   7.000Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                    4.186Ki        4.186Ki       +0.00%
¹ all samples are equal

                         │   a.txt    │                b.txt                │
                         │ allocs/op  │ allocs/op   vs base                 │
ComplexMessage_String-12   43.00 ± 0%   43.00 ± 0%       ~ (p=1.000 n=10) ¹
Lex-12                     29.00 ± 0%   29.00 ± 0%       ~ (p=1.000 n=10) ¹
Parse-12                   59.00 ± 0%   59.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    41.90        41.90       +0.00%
¹ all samples are equal
```